requirements.txt
# fastapi==0.109.0
# uvicorn[standard]==0.27.0
# python-multipart==0.0.6
# python-dotenv==1.0.0

# # File Processing
# PyMuPDF==1.23.0
# Pillow==10.2.0
# python-docx==1.1.0

# # AI & APIs
# openai==1.10.0
# google-api-python-client==2.116.0

# # AWS
# boto3==1.34.34

# # Database
# pymongo==4.6.1
# motor==3.3.2

# # Utils
# pydantic==2.5.0


# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Configuration
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.2

# File Processing
PyMuPDF==1.24.0
Pillow==11.0.0
python-docx==1.1.2

# AI & APIs
openai==1.54.0
google-api-python-client==2.150.0

# AWS
boto3==1.35.0

# Database (FIXED - Compatible versions)
pymongo==4.9.1
motor==3.6.0

# Additional
aiofiles==24.1.0
aiofiles>=23.2.1